Don’t error when dealing with requests with no user agent set.
The previous commit introduced a regression when dealing with requests
with no `User-Agent` header. This commit guards against that case.

According to the HTTP specification requests don’t have to have user
agents[1] so I think we should handle this case gracefully rather
than erroring. 

I also had to change the header encoding tests as MiniTest Mocks have
to be setup for each call and we are now calling `request.user_agent`
twice within that method.

[1] http://tools.ietf.org/html/rfc7231#section-5.5.3