Optionally use secure flag on cookie
For sites with PCI compliance requirements, setting insecure cookies
(i.e. ones that can be accessed in non-HTTPS contexts) causes
violations.

This adds a new configuration option `secure_cookie` that allows the
cookie to be set secure by default.

This doesn't (yet) apply to the vendored ahoy.js; I'll submit a similar
patch upstream to the ahoy JS project and filter it through that way.