Extract CookieJar from Tracker
Currently cookie-setting is baked into the Tracker class. I'd like to add a
new `Ahoy.secure_cookies` option to allow users to specify that cookies must
be sent over HTTPS only. Since this will represent an increase in complexity
of the cookie handling, I've decided to extract (and test) this in a separate
class which I can modify later to add the new option.