Remove jQuery dependency
Depending on jQuery is a big factor to not load Ahoy as part of the head of the
page, like most analytics library do.

However, in order to track page views and visits as soon as possible (so we
capture bounces effectively), we need it to run as soon as possible and so not
at the end of the body (after jQuery).

This change removes the dependency on jQuery by interacting with the browser
directly where necessary.

The change itself is based on a PR on the ahoy.js repository [1] and drops
support to IE8 and older browsers. The original source code for the vendorised
ahoy.js here is on our own fork of ahoy.js [2].

This version was tested on the following browsers via zuul + Sauce Labs:

browsers:
  - name: ie
    version: 9..latest
  - name: chrome
    version: latest
  - name: safari
    version: 9..latest
  - name: microsoftedge
    version: latest
  - name: iphone
    version: '9.0..latest'
  - name: android
    version: '5.0..latest'

[1]: https://github.com/ankane/ahoy.js/pull/10
[2]: https://github.com/Futurelearn/ahoy.js/tree/remove-jquery