Prevent encoding error due to invalid bytes
Trello: https://trello.com/c/GbKFWIYM

When posting a json body to /ahoy/visits with a referrer in the body
which includes invalid bytes, it raised with `Invalid byte sequence in
US-ASCII`.

The failure in our case was caused by:
"referrer" :
"https://search.naver.com/search.naver?where=nexearch&sm=tol_hty&query=%C7%BB%C3%B3%B7%B1"

There was already a `rescue nil` for errors occurring in the
`RefererParser`, but there are some sequences of characters that are
successfully parsed by the `RefererParser` but then fail on the
`presence` call, which does some string comparisons. This commit extends
the scope of the `rescue` to include errors raised by the call to
`presence`. This might mean that we mask other potential errors, but we
don't really use this search keyword anyway, so we'd prefer the code to
not error and return nil than fail and lose the visit tracking.

The test we wrote to reproduce the failure does not actually cause the
code to fail with an encoding error, because for some reason the URL
that fails on production raises an error in the `RefererParser` so it
is rescued by the existing `rescue nil`. However we thought we should
keep the test in as an illustration of what we're trying to achieve. We
spent some time trying to figure out why the same referrer was treated
differently when processed through the entire Rails stack versus
isolated testing, but couldn't come up with anything conclusive so we
parked the investigation in order to get the fix out.

We've verified that this works by `curl`ing an invalid referrer to the
`/ahoy/visits` endpoint.