Fix error when handling incorrect Content-Type and multiple events
This controller handles two types of requests:
1. single events (apparently a legacy API)
2. batch events

We occasionally have requests which were sent with either a missing
Content-Type header, or with the wrong value (application/x-www-form-urlencoded
instead of application/json). We think these are browser bugs.

The effect of `ActiveSupport::JSON.decode` compensates for this situation by
just assuming the request is JSON if the version parsed by Rails didn't contain
what was expected and parsing it manually.

Aside: it isn't clear why ActiveSupport::JSON.decode was added, but it was added
in: 94b14019bae67068338f9fd249ace38511bd9fa3

There's a problem though, because prior to this commit, the code "assumed" that
when manually parsing the JSON that it was a single event, when it can be either.
This would cause an error:

	TypeError: no implicit conversion of String into Integer

	[GEM_ROOT]/bundler/gems/ahoy-ec09a3179833/app/controllers/ahoy/events_controller.rb:26 :in `[]`
	[GEM_ROOT]/bundler/gems/ahoy-ec09a3179833/app/controllers/ahoy/events_controller.rb:26 :in `block in create`
	[GEM_ROOT]/bundler/gems/ahoy-ec09a3179833/app/controllers/ahoy/events_controller.rb:19 :in `each`
	[GEM_ROOT]/bundler/gems/ahoy-ec09a3179833/app/controllers/ahoy/events_controller.rb:19 :in `create`
	[GEM_ROOT]/gems/actionpack-5.0.5/lib/action_controller/metal/basic_implicit_render.rb:4 :in `send_action`
	[GEM_ROOT]/gems/actionpack-5.0.5/lib/abstract_controller/base.rb:188 :in `process_action`

I was able to reproduce the error by:
1. opening developer tools in a browser
2. triggering an _event_ request
3. Using "copy as cURL" to get a replayable/editable version of the request
4. Removing the Content-Type header or changing it to `application/x-www-form-urlencoded`
5. Running the tweaked command on the command line.

I would like to add tests, but there aren't any controller-level tests. There
is a "dead" PR to add them: https://github.com/ankane/ahoy/pull/56

https://trello.com/c/pkBJH15F/348-typeerror-no-implicit-conversion-of-string-into-integer-in-ahoy-events