Fix template errors when oversize event payloads are sent
Trello: https://trello.com/c/Shm5hd2O

Since upgrading to Rails 5.1, we've received occasional honeybadgers of
the following form:

    ActionView::MissingTemplate: Missing template ahoy/events/create,
    ahoy/base/create, application/create with {:locale=>[:en],
    :formats=>[:html, :text, :js, :css, :ics, :csv, :vcf, :png, :jpeg, :gif,
    :bmp, :tiff, :svg, :mpeg, :xml, :rss, :atom, :yaml, :multipart_form,
    :url_encoded_form, :json, :pdf, :zip, :gzip], :variants=>[],
    :handlers=>[:raw, :erb, :html, :builder, :ruby, :haml]}.

This error emitted from ahoy's handling of oversized payloads, which
responds with a 413 Payload Too Large response and a plain-text body.

It turns out that in Rails 5.1, the `render text:
'some-text'` option was removed in favour of `render plain:
'some-text'`[1]. In the former I believe the `text` option is now
completely ignored, so it's treated as a standard `render` call, and
goes looking for a template.

This updates the usage to the new style.

[1] https://guides.rubyonrails.org/layouts_and_rendering.html#rendering-text