Add Ahoy.preserve_callbacks option
Our Rails app uses a controller before_action to add additional
properties to ahoy events / visits, however these properties are not set
for any events fired by client side Javascript. On closer inspection this
is because although Ahoy::BaseController inherits from ApplicationController
it skips all controller before actions.

In order to ensure that this before_action runs we have introduced the
`Ahoy.preserve_callbacks`option which allows us to pass in an array of
callbacks which shouldn't be skipped.

It turns out this feature has been implemented in a more recent version
of ahoy [1] and so this shouldn't add to the complexity of upgrading
our fork of ahoy in the future.

[1] https://github.com/ankane/ahoy/commit/71129efe9d4d0b18a115db4787e812319eb99b8d

Co-authored-by: Seb Jacobs <me@sebjacobs.com>